Producto no se puede eliminar si plato esta en comanda y plato ya se puede crear pero solo con primera letra mayuscula

+++ src/main/java/org/springframework/samples/foorder/service/PlatoService.java
@@ -9,8 +9,12 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.foorder.model.Ingrediente;
 import org.springframework.samples.foorder.model.Plato;
+import org.springframework.samples.foorder.model.PlatoPedido;
 import org.springframework.samples.foorder.repository.IngredienteRepository;
+import org.springframework.samples.foorder.repository.PlatoPedidoRepository;
 import org.springframework.samples.foorder.repository.PlatoRepository;
+import org.springframework.samples.foorder.service.exceptions.PedidoPendienteException;
+import org.springframework.samples.foorder.service.exceptions.PlatoEnProcesoException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,12 +25,14 @@ public class PlatoService {
 	
 	private PlatoRepository platoRepository;
 	private IngredienteRepository ingredienteRepository;
+	private PlatoPedidoRepository platoPedidoRepository;
 	
 	@Autowired
-	public PlatoService(PlatoRepository platoRepository, IngredienteRepository ingredienteRepository) {
+	public PlatoService(PlatoRepository platoRepository, PlatoPedidoRepository platoPedidoRepository, IngredienteRepository ingredienteRepository) {
 		super();
 		this.platoRepository = platoRepository;
 		this.ingredienteRepository = ingredienteRepository;
+		this.platoPedidoRepository = platoPedidoRepository;
 	}
 
 	public Collection<Plato> findAll() {
@@ -43,10 +49,13 @@ public class PlatoService {
 		return platoRepository.save(plato);	
 	}
 	
-	@Transactional
-	public void deleteById(Integer id) {
+	@Transactional(rollbackFor = PlatoEnProcesoException.class)
+	public void deleteById(Integer id) throws PlatoEnProcesoException{
+		if(!platoSePuedeEliminar(id)) {
+			throw new PlatoEnProcesoException();
+		}
 		Plato plato = platoRepository.findById(id).get();
-		platoRepository.deleteById(id);	
+		platoRepository.deleteById(id);
 		log.info(String.format("Plate with name %s has been saved", plato.getName()));
 	}
 	
@@ -84,6 +93,20 @@ public class PlatoService {
 			}
 			if(falta) {
 				res.remove(plato);
+			}
+		}
+		return res;
+	}
+	
+	private Boolean platoSePuedeEliminar(Integer id) {
+		Boolean res = true;
+		Iterable<PlatoPedido> platosPedidos = platoPedidoRepository.findByPlatoId(id);
+		Iterator<PlatoPedido> pps = platosPedidos.iterator();
+		while (pps.hasNext()) {
+			PlatoPedido p = pps.next();
+			if(!(p.getEstadoplato().equals("FINALIZADO"))) {
+				res = false;
+				break;
 			}
 		}
 		return res;
